auth: allow signing on using a twitter screen name (DO NOT USE)
WARNING: This commit introduces a potential security breach. Use with
care in a controlled environment.

This makes the sign in code recognize your Twitter username and stores
the user id in the external uid field, which is used by other schemes
(LDAP). Security: if you set up a Twitter alias in your profile and
never use the sign in with Twitter button, then you are open to anyone
logging in as you provided:

1) You change your screen name in Twitter, but do not update your
profile in GitLab.
2) Someone else takes over your original screen name.

There is a window of time where an attack can be performed.

I AM USING THIS STRICTLY AS A PERSONAL CHANGE IN A CONTROLLED
ENVIRONMENT. DO NOT USE FOR PRODUCTION!

Signed-off-by: Alejandro Martinez Ruiz <alex@nowcomputing.com>